Fixing some error

+++ app/Http/Controllers/Author/PostController.php
@@ -48,6 +48,7 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+
         $this->validate($request, [
             'title'=>'required',
             'image' => 'required',
@@ -110,6 +111,10 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error("You are not authorized to access this post!", "Error");
+            return redirect()->back();
+        }
         return view('author.post.show', compact('post'));
     }
 
@@ -121,6 +126,10 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error("You are not authorized to access this post!", "Error");
+            return redirect()->back();
+        }
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit', compact('post','categories','tags'));
@@ -135,6 +144,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error("You are not authorized to access this post!", "Error");
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title'=>'required',
             'image' => 'image',
@@ -201,6 +215,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error("You are not authorized to access this post!", "Error");
+            return redirect()->back();
+        }
+
         if(Storage::disk('public')->exists('post/'.$post->image)){
             Storage::disk('public')->delete('post/'.$post->image);
         }
